pyproject.toml
# pyproject.toml

# Defines the build system and its requirements.
# 'setuptools' is a standard choice for building Python packages.
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

# Defines the core metadata for the project.
[project]
name = "can-logger"
version = "0.1.0"
description = "A lightweight, real-time CAN bus logger that saves specific signals to a JSON file."

# Lists the packages required for the project to run.
# uv or pip will install these when you set up the environment.
dependencies = [
    "python-can",
    "cantools",
]